requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-taggit==5.0.1
Markdown==3.5.2
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
